# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type ChatMessage struct {
	ID           pgtype.UUID        `json:"id"`
	RoomID       pgtype.UUID        `json:"room_id"`
	GameID       pgtype.UUID        `json:"game_id"`
	RoomPlayerID pgtype.UUID        `json:"room_player_id"`
	Message      string             `json:"message"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type Game struct {
	ID         pgtype.UUID        `json:"id"`
	RoomID     pgtype.UUID        `json:"room_id"`
	Status     string             `json:"status"`
	ConfigJson []byte             `json:"config_json"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	EndedAt    pgtype.Timestamptz `json:"ended_at"`
}

type GameEvent struct {
	ID           pgtype.UUID        `json:"id"`
	GameID       pgtype.UUID        `json:"game_id"`
	RoomPlayerID pgtype.UUID        `json:"room_player_id"`
	Type         string             `json:"type"`
	PayloadJson  []byte             `json:"payload_json"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type GamePlayer struct {
	ID           pgtype.UUID        `json:"id"`
	GameID       pgtype.UUID        `json:"game_id"`
	RoomPlayerID pgtype.UUID        `json:"room_player_id"`
	Role         pgtype.Text        `json:"role"`
	JoinedAt     pgtype.Timestamptz `json:"joined_at"`
	LeftAt       pgtype.Timestamptz `json:"left_at"`
}

type GameStateSnapshot struct {
	ID        pgtype.UUID        `json:"id"`
	GameID    pgtype.UUID        `json:"game_id"`
	Version   int32              `json:"version"`
	StateJson []byte             `json:"state_json"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Room struct {
	ID           pgtype.UUID        `json:"id"`
	Code         string             `json:"code"`
	PasswordHash pgtype.Text        `json:"password_hash"`
	SettingsJson []byte             `json:"settings_json"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type RoomPlayer struct {
	ID          pgtype.UUID        `json:"id"`
	RoomID      pgtype.UUID        `json:"room_id"`
	DisplayName string             `json:"display_name"`
	IsHost      bool               `json:"is_host"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UserID      pgtype.UUID        `json:"user_id"`
}

type User struct {
	ID           pgtype.UUID        `json:"id"`
	Email        string             `json:"email"`
	PasswordHash string             `json:"password_hash"`
	DisplayName  string             `json:"display_name"`
	AvatarUrl    pgtype.Text        `json:"avatar_url"`
	SettingsJson []byte             `json:"settings_json"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}